+++ bin/liveserver.dart
@@ -1,5 +1,0 @@
-import 'package:dliver/dliver.dart';
-
-main(){
-  var dliver = new DLiver();
-}