+++ ChangeLog.md
@@ -1,6 +1,17 @@
 ChangeLog
 =========
 
+1.0.3
+-----
+
+-	Added cli for dliver. pub global activate dliver - now worked;
+-	Added dependence from cupid
+
+1.0.2
+-----
+
+-	Fix DLiver class. Now it is so named: Dliver
+
 1.0.1
 -----